requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.1.78
locust==2.17.0
gunicorn==21.2.0

